Handle empty state with no scenarios
